asihpi - new checkpatch = more changes
    2d197673036728a8006148921209bd918d711c38
Mark hpetimer as BROKEN
    20863b72ef8bce4594c042ab5b5b8d0b4df8e568
Remove old Kconfig entry for pcsp
    d78b9225c26912a9636999f66b6488cbfca6cef6

+++ alsa/pci/asihpi/hpifunc.c
@@ -3668,14 +3668,14 @@ u16 HPI_FormatCreate(
 		break;
 	case HPI_FORMAT_PCM16_SIGNED:
 	case HPI_FORMAT_PCM16_BIGENDIAN:
-		Format.dwBitRate = (u32)wChannels *dwSampleRate * 2;
+		Format.dwBitRate = wChannels * dwSampleRate * 2;
 		break;
 	case HPI_FORMAT_PCM32_SIGNED:
 	case HPI_FORMAT_PCM32_FLOAT:
-		Format.dwBitRate = (u32)wChannels *dwSampleRate * 4;
+		Format.dwBitRate = wChannels * dwSampleRate * 4;
 		break;
 	case HPI_FORMAT_PCM8_UNSIGNED:
-		Format.dwBitRate = (u32)wChannels *dwSampleRate;
+		Format.dwBitRate = wChannels * dwSampleRate;
 		break;
 	default:
 		Format.dwBitRate = 0;
